requirements.txt
streamlit==1.37.1
Pillow==10.4.0
numpy==2.1.0


